+++ src/h/main.h
@@ -2,7 +2,7 @@
 #define MC_C_H
 
 #define MC_C_AUTHOR     "Author: Lily Awertnex"
-#define MC_C_NAME       "minecraft.c"
+#define MC_C_NAME       "heaven-hell_continuum"
 #define MC_C_VERSION    "0.2.0-alpha"
 
 #include "../engine/h/core.h"